Completed local permissions for deleteMenuItem().

+++ menuProject.py
@@ -428,6 +428,10 @@ def deleteMenuItem(restaurant_id, menu_id):
     restaurant = session.query(Restaurant).filter_by(id=restaurant_id).one()
     item = session.query(MenuItem).filter_by(id = menu_id).one()
 
+    # Check if user_id matches the user_id stored in login_session
+    if restaurant.user_id != login_session['user_id']:
+        return authorizationAlert("Delete")
+
     if request.method == 'POST':
         # delete restaurant object obtained from db query
         session.delete(item)
